+++ core/src/main/kotlin/org/pastalab/fray/core/observers/ScheduleRecorder.kt
@@ -3,33 +3,26 @@ package org.pastalab.fray.core.observers
 import java.io.File
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
+import org.pastalab.fray.core.ThreadContext
 import org.pastalab.fray.rmi.ScheduleObserver
 import org.pastalab.fray.rmi.TestStatusObserver
-import org.pastalab.fray.rmi.ThreadInfo
 
-class ScheduleRecorder : ScheduleObserver<ThreadInfo>, TestStatusObserver {
+class ScheduleRecorder : ScheduleObserver<ThreadContext>, TestStatusObserver {
   val recordings = mutableListOf<ScheduleRecording>()
 
   override fun onExecutionStart() {
     recordings.clear()
   }
 
-  override fun onNewSchedule(allThreads: List<ThreadInfo>, scheduled: ThreadInfo) {
-    var count = 0
-    var operation = ""
-    for (st in scheduled.stackTraces) {
-      if (st.className.startsWith("org.pastalab.fray")) {
-        continue
-      }
-      operation += "@${st.className}.${st.methodName},"
-      count += 1
-      if (count == 3) break
-    }
-    val enabled = allThreads.map { it.threadIndex }.toList()
-    val scheduledIndex = scheduled.threadIndex
-    val recording = ScheduleRecording(scheduledIndex, enabled, operation)
+  override fun onNewSchedule(allThreads: Collection<ThreadContext>, scheduled: ThreadContext) {
+    val enabled = allThreads.map { it.index }.toList()
+    val scheduledIndex = scheduled.index
+    val recording =
+        ScheduleRecording(scheduledIndex, enabled, scheduled.pendingOperation.javaClass.name)
     recordings.add(recording)
   }
+
+  override fun onContextSwitch(current: ThreadContext, next: ThreadContext) {}
 
   override fun onExecutionDone(bugFound: Throwable?) {}