+++ CHANGELOG.md
@@ -7,8 +7,11 @@
 - Allow Fray to use real-world time to decide if an operation should be unblocked.
 - Skip class constructors in `sun/security/ssl/SSLContextImpl`
 - Add instrumentation points at `NioSocket::accept`
+- Add thread pausing time logger.
 
 ### Changed
+
+- Implement more efficient schedule recorder/verifier.
 
 ### Deprecated